Add clone-extension command
Add a new command with which to clone an extension, optionally
with a username in the clone URL and copying in the Gerrit
commit hook.

Moves and extends the extension-directory finding code to the
base command class so that any command can use it. Adds
additional checks for the directory all the way up the file tree
and to the side at each level (e.g. it's possible to clone into
the extensions directory when you're somewhere down in includes).